Abstract the call functionality, and add TreeValueFunctionNode.
